+++ joodo/sample/src/sample/main.clj
@@ -2,6 +2,6 @@
 
 (defn -main []
   (let [port (Integer/parseInt (get (System/getenv) "PORT" "8080"))]
-    (joodo.tsukuri.GaeshiDevServer/main (into-array String ["-p" (str port) "-a" "127.0.0.1"]))))
+    (joodo.tsukuri.GaeshiDevServer/main (into-array String ["-p" (str port) "-a" "0.0.0.0"]))))